Move Util methods from rest.api.utils. to rest.api.common
