+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.common/src/main/resources/server-templates/formParams.mustache
@@ -1,0 +1,1 @@
+{{#isFormParam}}{{^isFile}}{{#isMultipart}}@Multipart{{/isMultipart}}{{^isMultipart}}@FormParam{{/isMultipart}}(value = "{{baseName}}"{{^required}}{{#isMultipart}}, required = false{{/isMultipart}}{{/required}})  {{{dataType}}} {{paramName}}{{/isFile}}{{#isFile}} @Multipart(value = "{{baseName}}"{{^required}}, required = false{{/required}}) InputStream {{paramName}}InputStream, @Multipart(value = "{{baseName}}" {{^required}}, required = false{{/required}}) Attachment {{paramName}}Detail{{/isFile}}{{/isFormParam}}
