+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.common/src/main/resources/server-templates/serviceFormParams.mustache
@@ -1,0 +1,1 @@
+{{#isFormParam}}{{^isFile}}{{{dataType}}} {{paramName}}{{/isFile}}{{#isFile}}InputStream {{paramName}}InputStream, Attachment {{paramName}}Detail{{/isFile}}{{/isFormParam}}
