+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.common/src/main/resources/service-catalog-api.yaml
