+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.util/src/main/java/org/wso2/carbon/apimgt/rest/api/util/utils/RestApiUtil.java
@@ -117,6 +117,17 @@ public class RestApiUtil {
     }
 
     /**
+     * To getting a message properties as a Map
+     * @param message - current inbound message
+     * @return Map object that contains all properties of cxf inbound message
+     */
+    public static HashMap<String,Object> addToJWTAuthenticationContext(Message message) {
+        HashMap<String,Object> hashMap = new HashMap<>();
+        message.forEach(hashMap::put);
+        return hashMap;
+    }
+
+    /**
      * Returns a generic errorDTO
      *
      * @param message specifies the error message
@@ -960,144 +971,6 @@ public class RestApiUtil {
         return returnedAPP;
     }
 
-    /**
-     * This is static method to return URI Templates map of API Store REST API.
-     * This content need to load only one time and keep it in memory as content will not change
-     * during runtime.
-     *
-     * @return URITemplate set associated with API Manager Store REST API
-     */
-    public static Set<URITemplate> getStoreAppResourceMapping(String version) {
-
-        API api = new API(new APIIdentifier(RestApiConstants.REST_API_PROVIDER,
-                RestApiConstants.REST_API_STORE_CONTEXT,
-                RestApiConstants.REST_API_STORE_VERSION_0));
-
-        if (storeResourceMappings != null) {
-            return storeResourceMappings;
-        } else {
-            try {
-                String definition;
-                if (RestApiConstants.REST_API_STORE_VERSION_0.equals(version)) {
-                    definition = IOUtils
-                            .toString(RestApiUtil.class.getResourceAsStream("/store-api.json"), "UTF-8");
-                } else {
-                    definition = IOUtils
-                            .toString(RestApiUtil.class.getResourceAsStream("/devportal-api.yaml"), "UTF-8");
-                }
-                APIDefinition oasParser = OASParserUtil.getOASParser(definition);
-                //Get URL templates from swagger content w created
-                storeResourceMappings = oasParser.getURITemplates(definition);
-            } catch (APIManagementException e) {
-                log.error("Error while reading resource mappings for API: " + api.getId().getApiName(), e);
-            } catch (IOException e) {
-                log.error("Error while reading the swagger definition for API: " + api.getId().getApiName(), e);
-            }
-            return storeResourceMappings;
-        }
-    }
-
-
-    /**
-     * This is static method to return URI Templates map of API Publisher REST API.
-     * This content need to load only one time and keep it in memory as content will not change
-     * during runtime.
-     *
-     * @return URITemplate set associated with API Manager publisher REST API
-     */
-    public static Set<URITemplate> getPublisherAppResourceMapping(String version) {
-        API api = new API(new APIIdentifier(RestApiConstants.REST_API_PROVIDER, RestApiConstants.REST_API_STORE_CONTEXT,
-                RestApiConstants.REST_API_STORE_VERSION_0));
-
-        if (publisherResourceMappings != null) {
-            return publisherResourceMappings;
-        } else {
-            try {
-                String definition;
-                if (RestApiConstants.REST_API_PUBLISHER_VERSION_0.equals(version)) {
-                    definition = IOUtils
-                            .toString(RestApiUtil.class.getResourceAsStream("/publisher-api.json"), "UTF-8");
-                } else {
-                    definition = IOUtils
-                            .toString(RestApiUtil.class.getResourceAsStream("/publisher-api.yaml"), "UTF-8");
-                }
-                APIDefinition oasParser = OASParserUtil.getOASParser(definition);
-                //Get URL templates from swagger content we created
-                publisherResourceMappings = oasParser.getURITemplates(definition);
-            } catch (APIManagementException e) {
-                log.error("Error while reading resource mappings for API: " + api.getId().getApiName(), e);
-            } catch (IOException e) {
-                log.error("Error while reading the swagger definition for API: " + api.getId().getApiName(), e);
-            }
-            return publisherResourceMappings;
-        }
-    }
-
-    /**
-     * This is static method to return URI Templates map of Service Catalog REST API.
-     * This content need to load only one time and keep it in memory as content will not change
-     * during runtime.
-     *
-     * @return URITemplate set associated with Service Catalog REST API
-     */
-    public static Set<URITemplate> getServiceCatalogAPIResourceMapping() {
-        API api = new API(new APIIdentifier(RestApiConstants.REST_API_PROVIDER,
-                RestApiConstants.REST_API_SERVICE_CATALOG_CONTEXT_FULL, "v0"));
-
-        if (serviceCatalogAPIResourceMappings != null) {
-            return serviceCatalogAPIResourceMappings;
-        } else {
-            try {
-                String definition;
-                definition = IOUtils
-                            .toString(RestApiUtil.class.getResourceAsStream("/service-catalog-api.yaml"), "UTF-8");
-                APIDefinition oasParser = OASParserUtil.getOASParser(definition);
-                //Get URL templates from swagger content we created
-                serviceCatalogAPIResourceMappings = oasParser.getURITemplates(definition);
-            } catch (APIManagementException e) {
-                log.error("Error while reading resource mappings for API: " + api.getId().getApiName(), e);
-            } catch (IOException e) {
-                log.error("Error while reading the swagger definition for API: " + api.getId().getApiName(), e);
-            }
-            return serviceCatalogAPIResourceMappings;
-        }
-    }
-
-    /**
-     * This is static method to return URI Templates map of API Admin REST API.
-     * This content need to load only one time and keep it in memory as content will not change
-     * during runtime.
-     *
-     * @return URITemplate set associated with API Manager Admin REST API
-     */
-    public static Set<URITemplate> getAdminAPIAppResourceMapping(String version) {
-
-        API api = new API(new APIIdentifier(RestApiConstants.REST_API_PROVIDER, RestApiConstants.REST_API_ADMIN_CONTEXT,
-                RestApiConstants.REST_API_ADMIN_VERSION_0));
-
-        if (adminAPIResourceMappings != null) {
-            return adminAPIResourceMappings;
-        } else {
-            try {
-                String definition;
-                if (RestApiConstants.REST_API_ADMIN_VERSION_0.equals(version)) {
-                    definition = IOUtils
-                            .toString(RestApiUtil.class.getResourceAsStream("/admin-api.json"), "UTF-8");
-                } else {
-                    definition = IOUtils
-                            .toString(RestApiUtil.class.getResourceAsStream("/admin-api.yaml"), "UTF-8");
-                }
-                APIDefinition oasParser = OASParserUtil.getOASParser(definition);
-                //Get URL templates from swagger content we created
-                adminAPIResourceMappings = oasParser.getURITemplates(definition);
-            } catch (APIManagementException e) {
-                log.error("Error while reading resource mappings for API: " + api.getId().getApiName(), e);
-            } catch (IOException e) {
-                log.error("Error while reading the swagger definition for API: " + api.getId().getApiName(), e);
-            }
-            return adminAPIResourceMappings;
-        }
-    }
 
     /**
      * Returns the white-listed URIs and associated HTTP methods for REST API by reading api-manager.xml configuration
@@ -1253,33 +1126,6 @@ public class RestApiUtil {
     public static boolean checkIfAnonymousAPI(Message inMessage) {
         return (inMessage.get(RestApiConstants.AUTHENTICATION_REQUIRED) != null &&
                 !((Boolean) inMessage.get(RestApiConstants.AUTHENTICATION_REQUIRED)));
-    }
-
-    /**
-     * This method is used to get the URI template set for the relevant REST API using the given base path.
-     *
-     * @param basePath Base path of the REST API
-     * @return Set of URI templates for the REST API
-     */
-    public static Set<URITemplate> getURITemplatesForBasePath(String basePath) {
-        Set<URITemplate> uriTemplates = new HashSet<>();
-        //get URI templates using the base path in the request
-        if (basePath.contains(RestApiConstants.REST_API_PUBLISHER_CONTEXT_FULL_0)) {
-            uriTemplates = RestApiUtil.getPublisherAppResourceMapping(RestApiConstants.REST_API_PUBLISHER_VERSION_0);
-        } else if (basePath.contains(RestApiConstants.REST_API_PUBLISHER_CONTEXT_FULL)) {
-            uriTemplates = RestApiUtil.getPublisherAppResourceMapping(RestApiConstants.REST_API_PUBLISHER_VERSION);
-        } else if (basePath.contains(RestApiConstants.REST_API_STORE_CONTEXT_FULL_0)) {
-            uriTemplates = RestApiUtil.getStoreAppResourceMapping(RestApiConstants.REST_API_STORE_VERSION_0);
-        } else if (basePath.contains(RestApiConstants.REST_API_DEVELOPER_PORTAL_CONTEXT_FULL)) {
-            uriTemplates = RestApiUtil.getStoreAppResourceMapping(RestApiConstants.REST_API_DEVELOPER_PORTAL_VERSION);
-        } else if (basePath.contains(RestApiConstants.REST_API_ADMIN_CONTEXT_FULL_0)) {
-            uriTemplates = RestApiUtil.getAdminAPIAppResourceMapping(RestApiConstants.REST_API_ADMIN_VERSION_0);
-        } else if (basePath.contains(RestApiConstants.REST_API_ADMIN_CONTEXT_FULL)) {
-            uriTemplates = RestApiUtil.getAdminAPIAppResourceMapping(RestApiConstants.REST_API_ADMIN_VERSION);
-        } else if (basePath.contains(RestApiConstants.REST_API_SERVICE_CATALOG_CONTEXT_FULL)) {
-            uriTemplates = RestApiUtil.getServiceCatalogAPIResourceMapping();
-        }
-        return uriTemplates;
     }
 
     /**
